+++ net-misc/unifi-controller-bin/unifi-controller-bin-5.12.22_rc.ebuild
@@ -1,13 +1,17 @@
+# Copyright 2019 Gentoo Authors
 # Distributed under the terms of the GNU General Public License v2
 
-EAPI=6
-CHECKREQS_DISK_VAR="500M"
+EAPI=7
+CHECKREQS_DISK_VAR="1G"
 
 inherit check-reqs unpacker user
 
+MY_HASH="478dfbf57c"
+#MY_DOC="372/2"
+
 MY_P="${P/-bin}"
 MY_PN="${PN/-bin}"
-MY_PV="${PV/_rc}"
+MY_PV="${PV/_rc}${MY_HASH:+-${MY_HASH}}"
 
 DESCRIPTION="Ubiquiti UniFi Controller"
 HOMEPAGE="https://www.ubnt.com/download/unifi/"
@@ -16,39 +20,56 @@ SRC_URI="
 	tools? (
 		https://dl.ubnt.com/unifi/${MY_PV}/unifi_sh_api -> unifi-${MY_PV}_api.sh
 	)"
+	#doc? (
+	#	https://community.ui.com/ubnt/attachments/ubnt/Blog_UniFi/${MY_DOC}/UniFi-changelog-5.10.x.txt -> unifi-${MY_PV}_changelog.txt
+	#)
 RESTRICT="mirror"
 
 LICENSE="GPL-3 UBNT-20170717"
 SLOT="0"
-KEYWORDS="amd64 arm x86"
-IUSE="nls rpi1 systemd +tools"
-UNIFI_LINGUAS=( ca cs da de_DE el en es_ES nl pl pt_PT sv tr zh_CN )
+KEYWORDS="~aarch64 ~amd64 ~arm ~x86"
+IUSE="nls rpi1 systemd +tools" # doc
+UNIFI_LINGUAS=( ca cs da de_DE el en es_ES fr ja nl pl pt_PT ru sv tr zh_CN zh_TW )
 IUSE+=" ${UNIFI_LINGUAS[@]/#/linguas_}"
 
 # debian control dependencies:
+#  adduser
 #  binutils
 #  coreutils
-#  jsvc
-#  mongodb-server (>=2.4.10) | mongodb-10gen (>=2.4.14) | mongodb-org-server (>=2.6.0)
-#  openjdk-7-jre-headless | java8-runtime-headless
-
+#  curl
+#  jsvc (>=1.0.8)
+#  libcap2
+#  mongodb-server (>= 2.4.10) | mongodb-10gen (>= 2.4.14) | mongodb-org-server (>= 2.6.0),
+#  mongodb-server (<< 1:3.6.0) | mongodb-10gen (<< 3.6.0) | mongodb-org-server (<< 3.6.0),
+#  java8-runtime-headless
+#
 # The version of mongodb bundled with the Mac edition is v2.4.14 at the moment,
 # but currently the oldest ebuild (and only v2.x) is v2.6.12.  The default
 # version is currently v3.0.14 - but this crashes with the UniFi code, possibly
 # documented in https://jira.mongodb.org/browse/SERVER-22334.
+#
 # As a result, we'll only accept the oldest or newer versions as dependencies.
+#
+# Ubiquiti recommend the use of MongoDB 3.4.x.
+#
+# ... which is now deprecated.  However, it is widely reported that the only issue with
+# MongoDB 3.6.x is that the '--nohttpinterface' option is now deprecated, and causes an
+# error if used.  The Ubiquiti code, of course, hard-codes this :(
+#
 DEPEND="
-	|| (
-		~dev-db/mongodb-2.6.12
-		>=dev-db/mongodb-3.2
-	)
-	>=virtual/jre-1.7.0
+	>=dev-db/mongodb-3.2
+	<dev-db/mongodb-3.7
+	>=virtual/jre-1.8.0
+	<virtual/jre-1.9.0
 "
 RDEPEND="${DEPEND}"
 
 S="${WORKDIR}"
 
-QA_PREBUILT="opt/${MY_P}/lib/native/*/*/libubnt_webrtc_jni.so"
+QA_PREBUILT="
+	opt/${MY_P}/lib/native/*/*/libubnt_sdnotify_jni.so
+	opt/${MY_P}/lib/native/*/*/libubnt_webrtc_jni.so
+"
 
 pkg_setup () {
 	# unifi controller uses mongodb as a data-store, and mongo immediately
@@ -79,16 +100,26 @@ src_unpack () {
 	done
 	cd "${S}"
 
-	if [[ "${ARCH}" == "arm" ]]; then
-		rm usr/lib/unifi/lib/native/Linux/x86_64/libubnt_webrtc_jni.so
-		use rpi1 && rm usr/lib/unifi/lib/native/Linux/armhf/libubnt_webrtc_jni.so
+	if [[ "${ARCH}" == "aarch64" ]]; then
+		rm usr/lib/unifi/lib/native/Linux/armv7/libubnt_{webrtc,sdnotify}_jni.so
+		rm usr/lib/unifi/lib/native/Linux/x86_64/libubnt_{webrtc,sdnotify}_jni.so
+	elif [[ "${ARCH}" == "arm" ]]; then
+		rm usr/lib/unifi/lib/native/Linux/aarch64/libubnt_{webrtc,sdnotify}_jni.so
+		use rpi1 && rm usr/lib/unifi/lib/native/Linux/armv7/libubnt_{webrtc,sdnotify}_jni.so
+		rm usr/lib/unifi/lib/native/Linux/x86_64/libubnt_{webrtc,sdnotify}_jni.so
 	elif [[ "${ARCH}" == "amd64" ]]; then
-		rm usr/lib/unifi/lib/native/Linux/armhf/libubnt_webrtc_jni.so
+		rm usr/lib/unifi/lib/native/Linux/aarch64/libubnt_{webrtc,sdnotify}_jni.so
+		rm usr/lib/unifi/lib/native/Linux/armv7/libubnt_{webrtc,sdnotify}_jni.so
 	else # [[ "${ARCH}" == "x86" ]]
-		rm usr/lib/unifi/lib/native/Linux/x86_64/libubnt_webrtc_jni.so
-		rm usr/lib/unifi/lib/native/Linux/armhf/libubnt_webrtc_jni.so
+		rm usr/lib/unifi/lib/native/Linux/aarch64/libubnt_{webrtc,sdnotify}_jni.so
+		rm usr/lib/unifi/lib/native/Linux/armv7/libubnt_{webrtc,sdnotify}_jni.so
+		rm usr/lib/unifi/lib/native/Linux/x86_64/libubnt_{webrtc,sdnotify}_jni.so
 	fi
-	rmdir -p usr/lib/unifi/lib/native/Linux/x86_64 usr/lib/unifi/lib/native/Linux/armhf 2>/dev/null
+	rmdir -p \
+		usr/lib/unifi/lib/native/Linux/aarch64 \
+		usr/lib/unifi/lib/native/Linux/armv7 \
+		usr/lib/unifi/lib/native/Linux/x86_64 \
+		2>/dev/null
 
 	rm -r usr/lib/unifi/lib/native/Windows
 	if [[ ${CHOST} == *-darwin* ]] ; then
@@ -102,13 +133,19 @@ src_unpack () {
 }
 
 src_prepare () {
+	local lingua=''
+
 	default
 
-	if use nls; then
-		local lingua=''
+	if use nls && (( ${#UNIFI_LINGUAS[@]} )); then
 		for lingua in ${UNIFI_LINGUAS[@]}; do
 			if ! use linguas_${lingua}; then
-				rm -r usr/lib/unifi/webapps/ROOT/app-unifi/locales/"${lingua}" || die
+				if [[ -d usr/lib/unifi/webapps/ROOT/app-unifi/locales/"${lingua}" ]]; then
+					rm -r usr/lib/unifi/webapps/ROOT/app-unifi/locales/"${lingua}" || die
+				fi
+				if [[ -d usr/lib/unifi/webapps/ROOT/app-unifi/data/locales/"${lingua}" ]]; then
+					rm -r usr/lib/unifi/webapps/ROOT/app-unifi/data/locales/"${lingua}" || die
+				fi
 			fi
 		done
 	fi
@@ -145,14 +182,17 @@ src_install () {
 	dosym /var/run/unifi /opt/"${MY_P}"/run
 
 	# <sigh>
-	dodir /opt/"${MY_P}"/bin
-	dosym /usr/bin/mongod /opt/"${MY_P}"/bin/mongod
+	#dodir /opt/"${MY_P}"/bin
+	#dosym /usr/bin/mongod /opt/"${MY_P}"/bin/mongod
+	exeinto "/opt/${MY_P}/bin"
+	newexe "${FILESDIR}"/mongod.sh mongod || die "Failed to install mongod wrapper"
 
 	if use tools; then
-		insinto /opt/"${MY_P}"/bin
-		newins "${WORKDIR}"/unifi-${MY_PV}_api.sh unifi-api.sh
-		fperms 755 /opt/"${MY_P}"/bin/unifi-api.sh
+		exeinto "/opt/${MY_P}/bin"
+		newexe "${WORKDIR}"/unifi-${MY_PV}_api.sh unifi-api.sh
 	fi
+
+	#use doc && newdoc "unifi-${MY_PV}_changelog.txt" "CHANGELOG-$(ver_cut '1-2').txt"
 
 	insinto /var/lib/unifi/data
 	doins "${FILESDIR}"/system.properties
@@ -195,6 +235,8 @@ pkg_postinst() {
 	elog "Additionally, ports 8881 and 8882 are reserved, and 6789 is used"
 	elog "for determining throughput."
 	elog
+	elog "From release 5.9.x onwards, port 8883/tcp must allow outbound traffic"
+	elog
 	elog "All of these ports may be customised by editing"
 	elog
 	elog "    /opt/${MY_P}/data/system.properties"
@@ -229,6 +271,9 @@ pkg_postinst() {
 	ewarn
 	ewarn "... in order to set appropriate Java XMS and XMX (minimum and"
 	ewarn "maximum memory constraints) values"
+	elog
+	ewarn "UniFi Controller 5.10+ requires at least firmware 4.0.9 for"
+	ewarn "UAP/USW and at least firmware 4.4.34 for USG"
 }
 
 pkg_prerm() {
